starting to add tests for arrays
