+++ pytealutils/abi/__init__.py
@@ -3,3 +3,16 @@ from .bytes import *
 from .collections import *
 from .ufixed import *
 from .uint import *
+
+
+def abiTypeName(t) -> str:
+    if hasattr(t, "__name__"):
+        return t.__name__.lower()
+    elif t.__origin__ is DynamicArray:
+        return "{}[]".format(abiTypeName(t.__args__[0]))
+    elif t.__origin__ is FixedArray:
+        return "{}[{}]".format(abiTypeName(t.__args__[0]), t.__args__[1])
+    elif t.__origin__ is Tuple:
+        return "({})".format(",".join([abiTypeName(a) for a in t.__args__]))
+
+    return ""
